chess-bot/requirements.txt
# Chess ML Bot - Dependencies

# Core chess library
python-chess>=1.999

# Machine learning framework
torch>=1.9.0
torchvision>=0.10.0

# Numerical computing
numpy>=1.21.0

# Optional: For GUI interface
pygame>=2.1.0

# Optional: For web interface
flask>=2.0.0

# Optional: For advanced analysis
scipy>=1.7.0
matplotlib>=3.5.0

# Optional: For tablebase support
# Note: Syzygy tablebases require separate download

# Development dependencies (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0